+++ pages/_app.tsx
@@ -31,7 +31,7 @@ import { LOCALSTORAGE_KEY } from '../analytics/common'
 import { mixpanelInit } from '../analytics/mixpanel'
 import { useLocalStorage } from '../helpers/useLocalStorage'
 import nextI18NextConfig from '../next-i18next.config.js'
-import { loadFeatureToggles } from '../helpers/useFeatureEnabled'
+import { loadFeatureToggles } from '../helpers/useFeatureToggle'
 
 function getLibrary(provider: any, connector: AbstractConnector | undefined): Web3 {
   const chainIdPromise = connector!.getChainId()
